+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/Converters.kt
@@ -513,7 +513,7 @@ internal inline fun realmObjectToRealmReferenceOrError(
 
 // Returns a converter fixed to convert objects of the given type in the context of the given mediator/realm
 internal fun <T> converter(
-    clazz: KClass<*>,
+    clazz: KClass<T & Any>,
     mediator: Mediator,
     realmReference: RealmReference
 ): RealmValueConverter<T> {