+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/query/ObjectQuery.kt
@@ -53,15 +53,8 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
     private val classKey: ClassKey,
     private val clazz: KClass<E>,
     private val mediator: Mediator,
-    composedQueryPointer: RealmQueryPointer? = null,
-    private val filter: String,
-    private vararg val args: Any?
+    internal val queryPointer: RealmQueryPointer,
 ) : RealmQuery<E>, InternalDeleteable, Thawable<Observable<RealmResultsImpl<E>, ResultsChange<E>>>, Flowable<ResultsChange<E>> {
-
-    internal val queryPointer: RealmQueryPointer = when {
-        composedQueryPointer != null -> composedQueryPointer
-        else -> parseQuery()
-    }
 
     private val resultsPointer: RealmResultsPointer by lazy {
         RealmInterop.realm_query_find_all(queryPointer)
@@ -70,7 +63,22 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
     private val classMetadata: ClassMetadata? = realmReference.schemaMetadata[clazz.simpleName!!]
 
     internal constructor(
-        composedQueryPointer: RealmQueryPointer?,
+        realmReference: RealmReference,
+        key: ClassKey,
+        clazz: KClass<E>,
+        mediator: Mediator,
+        filter: String,
+        args: Array<out Any?>
+    ) : this(
+        realmReference,
+        key,
+        clazz,
+        mediator,
+        parseQuery(realmReference, mediator, key, filter, args),
+    )
+
+    internal constructor(
+        composedQueryPointer: RealmQueryPointer,
         objectQuery: ObjectQuery<E>
     ) : this(
         objectQuery.realmReference,
@@ -78,25 +86,23 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         objectQuery.clazz,
         objectQuery.mediator,
         composedQueryPointer,
-        objectQuery.filter,
-        *objectQuery.args
     )
 
     override fun find(): RealmResults<E> =
         RealmResultsImpl(realmReference, resultsPointer, classKey, clazz, mediator)
 
-    override fun query(filter: String, vararg arguments: Any?): RealmQuery<E> {
-        return inputScope {
-            val appendedQuery = tryCatchCoreException {
-                RealmInterop.realm_query_append_query(
-                    queryPointer,
-                    filter,
-                    convertToQueryArgs(arguments, mediator, realmReference)
-                )
+    override fun query(filter: String, vararg arguments: Any?): RealmQuery<E> =
+        tryCatchCoreException {
+            inputScope {
+                val appendedQuery =
+                    RealmInterop.realm_query_append_query(
+                        queryPointer,
+                        filter,
+                        convertToQueryArgs(arguments, mediator, realmReference)
+                    )
+                ObjectQuery(appendedQuery, this@ObjectQuery)
             }
-            ObjectQuery(appendedQuery, this@ObjectQuery)
         }
-    }
 
     // TODO OPTIMIZE Descriptors are added using 'append_query', which requires an actual predicate.
     //  This might result into query strings like "TRUEPREDICATE AND TRUEPREDICATE SORT(...)". We
@@ -185,40 +191,45 @@ internal class ObjectQuery<E : BaseRealmObject> constructor(
         find().asInternalDeleteable().delete()
     }
 
-    private fun parseQuery(): RealmQueryPointer = tryCatchCoreException {
-        inputScope {
-            RealmInterop.realm_query_parse(
-                realmReference.dbPointer,
-                classKey,
-                filter,
-                convertToQueryArgs(args, mediator, realmReference)
-            )
-        }
-    }
-
-    private fun tryCatchCoreException(block: () -> RealmQueryPointer): RealmQueryPointer = try {
-        block.invoke()
-    } catch (exception: Throwable) {
-        throw CoreExceptionConverter.convertToPublicException(
-            exception,
-            customMessage = "Invalid syntax in query: ${exception.message}"
-        ) { coreException: RealmCoreException ->
-            when (coreException) {
-                is RealmCoreInvalidQueryStringException ->
-                    IllegalArgumentException("Wrong query string: ${coreException.message}")
-                is RealmCoreInvalidQueryException ->
-                    IllegalArgumentException("Wrong query field provided or malformed syntax in query: ${coreException.message}")
-                is RealmCoreIndexOutOfBoundsException ->
-                    IllegalArgumentException("Have you specified all parameters in your query?: ${coreException.message}")
-                else -> {
-                    // Use default mapping
-                    null
-                }
-            }
-        }
-    }
-
     override fun description(): String {
         return RealmInterop.realm_query_get_description(queryPointer)
+    }
+
+    companion object {
+        private fun parseQuery(
+            realmReference: RealmReference,
+            mediator: Mediator,
+            classKey: ClassKey,
+            filter: String,
+            args: Array<out Any?>
+        ): RealmQueryPointer = tryCatchCoreException {
+            inputScope {
+                val queryArgs = convertToQueryArgs(args, mediator, realmReference)
+                RealmInterop.realm_query_parse(realmReference.dbPointer, classKey, filter, queryArgs)
+            }
+        }
+
+        fun <R> tryCatchCoreException(block: () -> R): R =
+            try {
+                block.invoke()
+            } catch (exception: Throwable) {
+                throw CoreExceptionConverter.convertToPublicException(
+                    exception,
+                    customMessage = "Invalid syntax in query: ${exception.message}"
+                ) { coreException: RealmCoreException ->
+                    when (coreException) {
+                        is RealmCoreInvalidQueryStringException ->
+                            IllegalArgumentException("Wrong query string: ${coreException.message}")
+                        is RealmCoreInvalidQueryException ->
+                            IllegalArgumentException("Wrong query field provided or malformed syntax in query: ${coreException.message}")
+                        is RealmCoreIndexOutOfBoundsException ->
+                            IllegalArgumentException("Have you specified all parameters in your query?: ${coreException.message}")
+                        else -> {
+                            // Use default mapping
+                            null
+                        }
+                    }
+                }
+            }
     }
 }